Move classes into a classes folder
